move download gist logic to githubgistservice.
